# internal/adminapi/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "adminapi.ErrorResponse": {
                "properties": {
                    "error": {
                        "description": "Error is a short, human-readable explanation.",
                        "example": "rebuild failed: group \"x\": rules must not be empty",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "adminapi.QuarantineListResponse": {
                "properties": {
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/quarantine.Entry"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
...
            "quarantine.Entry": {
                "properties": {
                    "key": {
                        "type": "string"
                    },
                    "lastRetry": {
                        "type": "string"
                    },
                    "reason": {
                        "type": "string"
                    },
                    "retryCount": {
                        "type": "integer"
                    },
                    "section": {
                        "type": "string"
                    },
                    "since": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Send 'Bearer \u003ctoken\u003e'. The token is the contents of --admin-token-file on the server.",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "CRUD over the CRDT-backed sections of the policy: groups, facts, defaults, logging. The YAML loaded at boot is the floor; admin writes overlay it per-key and are gossiped to peer replicas. Every write rebuilds the effective config and either applies it atomically or quarantines the offending entry.",
        "title": "request-validator admin API",
        "version": "v1"
    },
...
    "paths": {
...
        "/api/v1/quarantine": {
            "get": {
                "description": "Per-node view: each replica may legitimately quarantine different items depending on its current state. Items are re-evaluated on every rebuild.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.QuarantineListResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List quarantined CRDT entries",
                "tags": [
                    "quarantine"
                ]
            }
        },
        "/api/v1/quarantine/{section}/{name}": {
            "delete": {
                "description": "The CRDT entry itself is NOT removed (it remains the source of truth); only the local quarantine buffer is cleared for this key.",
                "parameters": [
                    {
                        "description": "groups|facts|defaults|logging",
                        "in": "path",
                        "name": "section",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "key name (empty for singleton sections)",
                        "in": "path",
                        "name": "name",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Drop a quarantined entry without retry",
                "tags": [
                    "quarantine"
                ]
            }
        }
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/"
        }
    ]
}